refactor 채팅 코드 정리 및 css
채팅 코드 정리 및 css
