--- src/main/java/com/hermez/farrot/chat/chatmessage/controller/ConnectRequest.java
@@ -1,5 +1,0 @@
-package com.hermez.farrot.chat.chatmessage.controller;
-
-public record ConnectRequest(String userEmail) {
-
-}
