+++ src/main/java/com/hermez/farrot/chat/chatmessage/dto/request/ConnectRequest.java
@@ -1,0 +1,5 @@
+package com.hermez.farrot.chat.chatmessage.dto.request;
+
+public record ConnectRequest(String userEmail) {
+
+}
