+++ src/main/java/com/hermez/farrot/chat/chatmessage/dto/response/ConnectResponse.java
@@ -1,4 +1,4 @@
-package com.hermez.farrot.chat.chatmessage.controller;
+package com.hermez.farrot.chat.chatmessage.dto.response;
 
 public record ConnectResponse(String type,String userEmail) {
