+++ src/main/java/com/hermez/farrot/chat/chatroom/dto/response/ReadCountResponse.java
@@ -1,5 +1,4 @@
 package com.hermez.farrot.chat.chatroom.dto.response;
 
 public record ReadCountResponse(Integer readCount) {
-
 }
