+++ src/main/resources/static/css/chat/chat.css
@@ -1,0 +1,158 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+html, body {
+  height: 100%;
+}
+
+.wrapper {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+
+.user-container label {
+  font-size: 12px;
+  margin-right: 1rem;
+}
+
+.user-container input {
+  border-radius: 3px;
+  border: none;
+  height: 100%;
+}
+
+.display-container {
+  flex: 12;
+  background: #b2c7d9;
+  overflow-y: scroll;
+}
+
+.input-container {
+  flex: 1;
+  display: flex;
+  justify-content: stretch;
+  align-items: stretch;
+}
+
+.input-container span {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 0.3rem;
+  width: 100%;
+}
+.chatting-input{
+  font-size: 12px;
+  height: 100%;
+  flex: 8;
+  border: none;
+}
+.send-button{
+  flex: 1;
+  background: #3991c6;
+  border: none;
+  height: 100%;
+  border-radius: 3px;
+}
+.chatting-list li{
+  width: 90%;
+  padding: 0.3rem;
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-end;
+  margin-top: 0.5rem;
+}
+.profile{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex: 1;
+}
+.profile.user{
+  font-size: 10px;
+  margin-bottom: 0.3rem;
+}
+.message{
+  border-radius: 5px;
+  padding: 0.5rem;
+  font-size: 12px;
+  margin: 0.5px;
+  flex: 7;
+}
+.message-img{
+  object-fit: cover;
+  max-height: 300px;
+  max-width: 200px;
+}
+.read-count{
+    font-size: 10px;
+    margin: 0 5px;
+}
+.time{
+  font-size: 10px;
+  margin: 0 5px;
+}
+.sent{
+  flex-direction: row-reverse;
+  float: right;
+}
+.sent .message{
+  background: #3991c6;
+}
+.received .message{
+  background: #fff;
+}
+.message-img{
+  width: auto;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: 0.3s;
+}
+.message-img:hover{opacity: 0.8;}
+.modal{
+  display: none;
+  position: fixed;
+  z-index: 1;
+  padding-top: 100px;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  cursor: pointer;
+  background-color: rgba(0,0,0,0.8);
+}
+.modal_content{
+  margin: auto;
+  display: block;
+  width: 50%; height: auto;
+  max-width: 1000px;
+  border-radius: 10px;
+  animation-name: zoom;
+  animation-duration: 0.8s;
+}
+@keyframes zoom {
+  from{transform: scale(0)}
+  to{transform: scale(1)}
+}
+.close{
+  position: absolute;
+  top: 15px;
+  right: 35px;
+  color: #f1f1f1;
+  font-size: 40px;
+  font-weight: bold;
+  transition: 0.3s;
+}
+.close:hover, .close:focus{
+  color: #bbb;
+  text-decoration: none;
+  cursor: pointer;
+}
