+++ src/main/resources/templates/chat/chat-room.html
@@ -5,7 +5,7 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.5.2/sockjs.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"></script>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-  <link th:href="@{/css/chat.css}" rel="stylesheet">
+  <link th:href="@{/css/chat/chat.css}" rel="stylesheet">
   <style>
     #response {
       height: 500px;
